chore(ORCID): Integrate orcid codec
